Implement Song Importing
This commit implements functionality for users to be able to select a
folder containing songs and have them be imported into the program. The
user can now pick and choose which songs they want to be imported
from a selected folder, as well as edit metadata of imported songs.
The user also now has the ability to remove or edit songs in the main
song list.

Other minor changes:
- Switching from qdoc to doxygen for documentation.
- Switching from using QMediaPlayer for metadata handling to using
    Taglib. The Readme has been updated to give attribution and specify
    which version is being used.